+++ source/app/layout/nodepositions.h
@@ -42,7 +42,7 @@ private:
     mutable std::thread::id _threadId;
 
     float _scale = 1.0f;
-    int _smoothing = 1;
+    size_t _smoothing = 1;
 
     QVector3D getNoLocking(NodeId nodeId) const;
 
@@ -59,8 +59,8 @@ public:
     void setScale(float scale) { _scale = scale; }
     float scale() const { return _scale; }
 
-    void setSmoothing(int smoothing) { Q_ASSERT(smoothing <= MAX_SMOOTHING); _smoothing = smoothing; }
-    int smoothing() const { return _smoothing; }
+    void setSmoothing(size_t smoothing) { Q_ASSERT(smoothing <= MAX_SMOOTHING); _smoothing = smoothing; }
+    size_t smoothing() const { return _smoothing; }
 
     QVector3D get(NodeId nodeId) const;
     std::vector<QVector3D> get(const std::vector<NodeId>& nodeIds) const;
